# Optimize performance of amp