<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
Others
### Describe
<!-- Describe what this PR does -->
Optimize the performance of `unscale` function when using AMP.

Unit: samples/s
| model      | before opt | after opt          | speedup
| ---------- | ---------- | ------------------ |--|
| bert_large |   51.87563         | 56.37118           | 1.09x